+++ Classes/Command/CleanupChatsCommand.php
@@ -1,8 +1,9 @@
 <?php
+declare(strict_types=1);
 /**
- * Class CleanupCommandController
+ * Class CleanupChatsCommand
  *
- * Copyright (C) Leipzig University Library 2022 <info@ub.uni-leipzig.de>
+ * Copyright (C) Leipzig University Library 2024 <info@ub.uni-leipzig.de>
  *
  * @author  Frank Morgner <morgnerf@ub.uni-leipzig.de>
  * @license http://opensource.org/licenses/gpl-2.0.php GNU General Public License
@@ -20,31 +21,25 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  */
-
 namespace Ubl\Supportchat\Command;
 
-use TYPO3\CMS\Core\Database\Connection;
+use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Input\InputOption;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Style\SymfonyStyle;
 use TYPO3\CMS\Core\Database\ConnectionPool;
-use \TYPO3\CMS\Core\Utility\GeneralUtility;
-use \TYPO3\CMS\Extbase\Mvc\Controller\CommandController;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
 
 /**
- * Class ChatCleanupCommandController
+ * Class CleanupCommandController
  *
  * Provides commandline interface to clean up past bookings
  *
  * @package Ubl\Booking\Command
  */
-class CleanupCommandController extends CommandController
+class CleanupChatsCommand extends Command
 {
-    /**
-     * The object manager
-     *
-     * @var \TYPO3\CMS\Extbase\Object\ObjectManager
-     * @inject
-     */
-    protected $objectManager;
-
     /**
      * Size of chunk for large scales sql queries
      *
@@ -54,26 +49,50 @@ class CleanupCommandController extends CommandController
     protected $sqlOperatingChunksize = 250;
 
     /**
-     * Constructor
+     * Configure the command by defining the name, options and arguments
      *
      * @return void
      */
-    public function __construct()
+    public function configure() :void
     {
-        if (!$this->objectManager) {
-            $this->objectManager = GeneralUtility::makeInstance('TYPO3\CMS\Extbase\Object\ObjectManager');
-        }
+        $this
+            ->setDescription('Removes passed chats after a certain period of days.')
+            ->setHelp('Removes passed chats after a certain period of days.'
+                . LF . 'To specify the period of days chat will be kept, use the --days or -d option. Default is 7 days.'
+                . LF . 'To see only data what will be removed, use the --dry-run option.'
+            )
+            ->addOption(
+                'days',
+                'd',
+                InputOption::VALUE_REQUIRED,
+                'Pass this option to set time interval defined in days for cleaning frontend user table. Default is 7.',
+                '7'
+            )->addOption(
+                'dry-run',
+                null,
+                InputOption::VALUE_NONE,
+                'If this option is set, the files will not be processed.'
+            );
     }
 
     /**
-     * Clean up and removes user from frontend user table to prevent collecting an amount of trusted data
+     * Console command for removing expired bookings
      *
-     * @param int $days 	How long user in fe_users should be kept. Default is 7 days.
-     *
-     * @return void
+     * @param InputInterface $input
+     * @param OutputInterface $output
+     * @return int
      */
-    public function cleanupChatMessagesCommand($days = 7)
+    protected function execute(InputInterface $input, OutputInterface $output): int
     {
+        $isDryRun = $input->getOption('dry-run') != false ? true : false;
+
+        $io = new SymfonyStyle($input, $output);
+        $io->title($this->getDescription());
+        if ($isDryRun === true) {
+            $io->writeln('<info>This is a dry-run. Data will not be removed.</info>');
+        }
+        $days = $input->getOption('days');
+
         $dt = new \DateTimeImmutable('now', new \DateTimeZone(date_default_timezone_get()));
         $current = $dt->modify('midnight');
         $ts = $current->sub(new \DateInterval("P{$days}D"));
@@ -83,29 +102,19 @@ class CleanupCommandController extends CommandController
             $this->sqlOperatingChunksize
         );
         $cnt = 0;
-        foreach ($uids as $package) {
-            $cnt += $this->removeChatMessagesByIds($package);
-            sleep(3);
+        if ($isDryRun === false) {
+            foreach ($uids as $package) {
+                $cnt += $this->removeChatMessagesByIds($package);
+                sleep(3);
+            }
+        } else {
+            $cnt = count($chatList);
         }
-        $this->outputLine('%d chats removed before %s', [$cnt, $ts->format('d.m.Y H:i:s')]);
+        $io->writeln(
+            sprintf('%d chats removed before %s', $cnt, $ts->format('d.m.Y H:i:s'))
+        );
+        return 0;
     }
-
-    /**
-     * Clean up and removes user from frontend user table to prevent collecting an amount of trusted data
-     *
-     * @param int $days 	How long user in fe_users should be kept. Default is 30 days.
-     *
-     * @return void
-     */
-    public function cleanupChatLogsCommand($days = 30)
-    {
-        $dt = new \DateTimeImmutable('now', new \DateTimeZone(date_default_timezone_get()));
-        $current = $dt->modify('midnight');
-        $ts = $current->sub(new \DateInterval("P{$days}D"));
-        $cnt = $this->removeChatLogsBeforeTime($ts);
-        $this->outputLine('%d chats removed before %s', [$cnt, $ts->format('d.m.Y H:i:s')]);
-    }
-
 
     /**
      * Get connection for table
@@ -187,33 +196,7 @@ class CleanupCommandController extends CommandController
                     $queryBuilder->expr()->in('chat_pid', $deleteList)
                 )
                 ->execute();
-        } catch (Exception $e) {
-            'Error while operating on database:' . $e->getMessage() . ' with code:' . $e->getCode();
-        }
-    }
-
-    /**
-     * Remove chat logs before a specified time
-     *
-     * @param \DateTimeInterface $time Timestamp
-     *
-     * @return int 	Affected row to proceed.
-     * @access public
-     */
-    public function removeChatLogsBeforeTime(\DateTimeInterface $time)
-    {
-        try {
-            $queryBuilder = $this->getConnectionForTable('tx_supportchat_log');
-            return $queryBuilder
-                ->delete('tx_supportchat_log')
-                ->where(
-                    $queryBuilder->expr()->lt(
-                        'tstamp',
-                        $time->getTimestamp()
-                    )
-                )
-                ->execute();
-        } catch (Exception $e) {
+        } catch (\Exception $e) {
             'Error while operating on database:' . $e->getMessage() . ' with code:' . $e->getCode();
         }
     }